feat(GN-69): 로그인 페이지 초기 레이아웃 작업
