+++ package.json
@@ -43,7 +43,8 @@
     "next": "14.2.5",
     "react": "^18",
     "react-datepicker": "^7.3.0",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-hook-form": "^7.52.1"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.3.0",
